+++ integrations/github/github/core/exporters/pull_request_exporter.py
@@ -1,3 +1,5 @@
+from typing import Any, Dict
+from github.clients.rest_client import GithubRestClient
 from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE, RAW_ITEM
 from loguru import logger
 from github.core.options import SinglePullRequestOptions, ListPullRequestOptions
@@ -5,7 +7,7 @@ from github.clients.base_client import AbstractGithubClient
 from github.core.exporters.abstract_exporter import AbstractGithubExporter
 
 
-class PullRequestExporter(AbstractGithubExporter[AbstractGithubClient]):
+class RestPullRequestExporter(AbstractGithubExporter[GithubRestClient]):
 
     async def get_resource[
         ExporterOptionsT: SinglePullRequestOptions
@@ -13,23 +15,26 @@ class PullRequestExporter(AbstractGithubExporter[AbstractGithubClient]):
         repo_name = options["repo_name"]
         pr_number = options["pr_number"]
 
-        endpoint = f"repos/{self.client.organization}/{repo_name}/pulls/{pr_number}"
+        endpoint = (
+            f"{self.client.base_url}/repos/{self.client.organization}/{repo_name}/pulls/{pr_number}"
+        )
         response = await self.client.send_api_request(endpoint)
 
         logger.debug(f"Fetched pull request with identifier: {repo_name}/{pr_number}")
 
-        return response.json()
+        return response
 
     async def get_paginated_resources[
         ExporterOptionsT: ListPullRequestOptions
     ](self, options: ExporterOptionsT,) -> ASYNC_GENERATOR_RESYNC_TYPE:
         """Get all pull requests in the organization's repositories with pagination."""
-
-        repo_name = options["repo_name"]
+        
+        params: Dict[str, Any] = dict(options) if options else {}
+        repo_name = params.pop("repo_name")
 
         async for pull_requests in self.client.send_paginated_request(
-            f"repos/{self.client.organization}/{repo_name}/pulls",
-            {"state": options["state"]},
+            f"{self.client.base_url}/repos/{self.client.organization}/{repo_name}/pulls",
+            params,
         ):
             logger.info(
                 f"Fetched batch of {len(pull_requests)} pull requests from repository {repo_name}"