+++ integrations/github/github/helpers/utils.py
@@ -8,3 +8,5 @@ class GithubClientType(StrEnum):
 
 class ObjectKind(StrEnum):
     REPOSITORY = "repository"
+    PULL_REQUEST = "pull-request"
+