+++ integrations/github/github/webhook/webhook_processors/pull_request_webhook_processor.py
@@ -1,7 +1,7 @@
 from typing import cast
 from loguru import logger
 from github.webhook.events import PULL_REQUEST_EVENTS
-from github.utils import ObjectKind, PullRequestState
+from github.helpers.utils import ObjectKind
 from github.clients.client_factory import create_github_client
 from github.webhook.webhook_processors.github_abstract_webhook_processor import (
     _GithubAbstractWebhookProcessor,
@@ -12,16 +12,26 @@ from port_ocean.core.handlers.webhook.webhook_event import (
     WebhookEvent,
     WebhookEventRawResults,
 )
-from github.core.exporters.pull_request_exporter import PullRequestExporter
+from github.core.exporters.pull_request_exporter import RestPullRequestExporter
 from github.core.options import SinglePullRequestOptions
 from integration import GithubPullRequestConfig
+from github.webhook.webhook_processors.base_repository_webhook_processor import (
+    BaseRepositoryWebhookProcessor,
+)
 
 
-class PullRequestWebhookProcessor(_GithubAbstractWebhookProcessor):
+class PullRequestWebhookProcessor(BaseRepositoryWebhookProcessor):
+
+    async def _validate_payload(self, payload: EventPayload) -> bool:
+        return (
+            "pull_request" in payload
+            and "number" in payload["pull_request"]
+        )
+        
     async def _should_process_event(self, event: WebhookEvent) -> bool:
         return (
             event.headers.get("x-github-event") == "pull_request"
-            and event.payload["action"] in PULL_REQUEST_EVENTS
+            and event.payload.get("action") in PULL_REQUEST_EVENTS
         )
 
     async def get_matching_kinds(self, event: WebhookEvent) -> list[str]:
@@ -38,27 +48,17 @@ class PullRequestWebhookProcessor(_GithubAbstractWebhookProcessor):
         logger.info(f"Processing pull request event: {action} for {repo_name}/{number}")
 
         config = cast(GithubPullRequestConfig, resource_config)
-        if action == "closed" and config.selector.state == PullRequestState.OPEN:
-            logger.info(f"Pull request {repo_name}/{number} was closed")
-
+        if action == "closed" and config.selector.state == "open":
             return WebhookEventRawResults(
                 updated_raw_results=[],
                 deleted_raw_results=[pull_request],
             )
 
-        exporter = PullRequestExporter(create_github_client())
+        exporter = RestPullRequestExporter(create_github_client())
         data_to_upsert = await exporter.get_resource(
             SinglePullRequestOptions(repo_name=repo_name, pr_number=number)
         )
 
         return WebhookEventRawResults(
             updated_raw_results=[data_to_upsert], deleted_raw_results=[]
-        )
-
-    async def validate_payload(self, payload: EventPayload) -> bool:
-        if not {"action", "pull_request", "repository"} <= payload.keys():
-            return False
-
-        return bool(
-            payload["repository"].get("name") and payload["pull_request"].get("number")
         )